Fix unescaped "<" in MathML in PDF (#18520)
Link to issue number:

Fixes #18511
Summary of the issue:

If the structure tree is used for MathML in PDF, calls to getValue return the interpreted &gt; so that it becomes < and we have an HTML syntax error.
Description of user facing changes:

The bug is fixed.
Description of developer facing changes:

None.
Description of development approach:

Called html.escape() for the result of the getValue (when it was not "none")