Use GitHub forms for issues (#15320)
### Summary of the issue:

[GitHub issue
forms](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/about-issue-and-pull-request-templates#issue-templates)
provide an easier and more clear way for users to report issues on the
NVDA project.

Currently, when filling out an issue form, inline comments and an
external guide is required to understand how to fill out an issue.

### Description of user facing changes

When reporting an issue to NVDA, users can use the simpler GitHub forms
interface.
Information from inline comments and the [issue
guide](https://github.com/nvaccess/nvda/blob/master/devDocs/githubIssueTemplateExplanationAndExamples.md)
have been merged into this form, so that it is easier to understand the
requirements.

The old forms have been retained and simplified for regular issue
reporters who prefer these.

### Description of development approach

N/A

### Testing strategy:

These changes have been pushed to a test repository

The new issue interface can be tested here:
https://github.com/nvaccess/nvda-issue-form-test/issues/new/choose

Example issues filled out using the form are as follows:
- https://github.com/nvaccess/nvda-issue-form-test/issues/1
- https://github.com/nvaccess/nvda-issue-form-test/issues/2

### Known issues with pull request:

The readability of markdown blocks might not be ideal, refer to
https://github.com/orgs/community/discussions/61696.
Tabbing through markdown editors is overly verbose and a frustrating
experience.