Modernize installer and reg key usage (#18851)
Part of #16304
Summary of the issue:

We need to refactor large parts of the installer code as part of the 64bit migration.
This is to ensure the old copy of NVDA is cleaned up correctly.
Description of user facing changes:

None
Description of developer facing changes:

    Some API breaking changes and deprecations which shouldn't affect most/any add-ons

Description of development approach:

    Move relevant keys to config.registry
    Move relevant paths to WritePaths
    Add type hints
    Create a winBinding for RegDeleteTreeW to update _deleteKeyAndSubkeys. The current design of _deleteKeyAndSubkeys was to support Windows XP and older