Fix bug when updating add-ons automatically (#18971)
Link to issue number:

Fixes #18965
Summary of the issue:

When NVDA is configured to update add-ons automatically without provide notifications, add-ons aren't updated and an error is produced.
Description of user facing changes:

NVDA can be configured to update add-ons automatically, and this should work as expected.
Description of developer facing changes:

None.
Description of development approach:

Use wx.CallAfter to present a message informing that add-ons are been updated, in the _checkForUpdatableAddonsmethod of the UpdatableAddonsDialog class, just after the condition to check that add-ons should be updated automatically, ensuring that this can be run in the main thread.

Additionally, an error has been discovered and fixed in addonHandler, when installing an add-on which shows a message before installation.