Return local control when local machine locks (#18973)
Link to issue number:

Fixes #18951
Summary of the issue:

When connected as leader, if controling the remote computer, the local computer locking results in the user being unable to use their device, as NVDA is still capturing keys at the lock screen.
Description of user facing changes:

If controlling a remote computer, control is returned to the local computer when the local computer is locked. If this has happened, NVDA Remote Access will resume controlling the remote computer when the local computer is unlocked.
Description of developer facing changes:

None
Description of development approach:

Add a post_sessionLockStateChanged handler to _remoteClient.client.Client. The handler is registered upon successful connection as leader, and unregistered when disconnecting as leader. The handler:

    Upon the device being locked, if controlling the remote computer, switches to local control, and sets a flag indicating that it has done so.
    Upon the device being unlocked, if the aforesaid flag is set, returns to controlling the remote computer, and unsets the flag.

Testing strategy:

    Ran NVDA from source. Connected as leader to a channel with a follower.
    Switched to controlling the remote computer. Waited for my device to lock as part of its power plan. Ensured that control was returned to the local computer.
    Unlocked my computer. Ensured control was returned to the remote computer.
    Returned control to the local computer, and locked my computer. Ensured that control was not moved.
    Unlocked my computer. Ensured that control was not moved.