Added ability to report spelling errors in braille (#18641)
Fixes #7608

### Summary of the issue:
Currently, spelling errors aren't presented in braille. This may
negatively impact deaf-blind people and other braille users.

### Description of user facing changes:
If NVDA is configured to report format changes with tags in braille, and
to report spelling errors in braille, these mistakes would be presented
in braille with the "e" and "e with dot 7" tag.
An unassigned command has been added to toggle the reporting of spelling
errors in braille. This can also be configured from the Document
Formatting dialog.

### Description of developer facing changes:
None.

### Description of development approach:
Added a new item in the `fontAttributeFormattingMarkers` dictionary of
`braille.py` for spelling errors (`invalid-spelling`).
A `shouldBeUsed` method has been added to the `FormattingMarker`class,
to determine if an attribute should be reported.

### Testing strategy:
Tested locally in Microsoft Word and Notepad, typing text detected as
spelling errors.

### Known issues with pull request:
None.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Cyrille Bougot <cyrille.bougot2@laposte.net>
Co-authored-by: Sascha Cowley <16543535+SaschaCowley@users.noreply.github.com>