Fix x64 identification for update check #19020
Fixes #18983
Summary of the issue:

NVDA incorrectly identified the OS architecture as 32bit when running as a 64bit process.
This caused the update check to return 2025.3 as the expected NVDA version, as NVDA falsely said the system was 32bit.
Description of user facing changes:

Update check works for alphas
Description of developer facing changes:

None
Description of development approach:

Fix PROCESSOR_ARCHITEW6432 usages