Avoid runing AI image descriptions while screen curtain is enabled (#19057)
Fixes #19045
Summary of the issue:

Image Description not disabled while Screen Curtain is enabled
Description of user facing changes:

Users will be notified that AI image description is unavailable when they try to use it while the screen curtain is enabled.
Description of developer facing changes:

The item of SCREEN_CURTAIN is added to gui.blockAction.Context , developers can use it in decoration to avoid actions when screen curtain is enabled.