Fix liblouis access violation errors (#19070)
Fixes #19025
Summary of the issue:

Translation can cause access violations on 64 bit version of NVDA.
Description of user facing changes:

Errors are hopefully fixed
Description of developer facing changes:

None
Description of development approach:

Our table resolver was returning an array of tables, but that array was not terminated with a NULLPTR as expected by liblouis. Added the None/NULLPTR to the end of the array.