Fix xliff regeneration (#19092)
### Link to issue number:
<!-- Use Closes/Fixes/Resolves #xxx to link this PR to the issue it is
responding to. -->
Fixes #19023
Replaces pr #19054 

### Summary of the issue:
The GitHub action that regenerates updated xliff files for translation,
from English user documentation is currently broken, as the ssh
credentials uses an out-of-date key.

### Description of user facing changes:

### Description of developer facing changes:

### Description of development approach:
The GitHub action now uses a GitHub Personal Access Token (PAT) that has
permission to push to the repository, and will still trigger future
actions.

### Testing strategy:
* [x] Run and ensure it successfully pushes to the pr branch, by
temporarily listing the pr branch for the action.
 * [ ] Once merged to beta, ensure the action correctly runs.

### Known issues with pull request:
None known.

### Code Review Checklist:

<!--
This checklist is a reminder of things commonly forgotten in a new PR.
Authors, please do a self-review of this pull-request.
Check items to confirm you have thought about the relevance of the item.
Where items are missing (eg unit / system tests), please explain in the
PR.
To check an item `- [ ]` becomes `- [x]`, note spacing.
You can also check the checkboxes after the PR is created.
A detailed explanation of this checklist is available here:

https://github.com/nvaccess/nvda/blob/master/projectDocs/dev/githubPullRequestTemplateExplanationAndExamples.md#code-review-checklist
-->

- [x] Documentation:
  - Change log entry
  - User Documentation
  - Developer / Technical Documentation
  - Context sensitive help for GUI changes
- [x] Testing:
  - Unit tests
  - System (end to end) tests
  - Manual testing
- [x] UX of all users considered:
  - Speech
  - Braille
  - Low Vision
  - Different web browsers
  - Localization in other languages / culture than English
- [x] API is compatible with existing add-ons.
- [x] Security precautions taken.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: GitHub Actions <actions@github.com>