Update pyright to use [nodejs] (#19078)
Closes #17749
Summary of the issue:

#17744 introduced pyright using it's default server nodeenv.
The pyright python package recommends installing pyright[nodejs] for a faster and more reliable server.
Hwever we couldn't use it while NVDA was 32bit.
This is because the node wheel binaries required a 64bit environment.
Now that NVDA has migrated to 64bit, we should update the pyright package version to using nodejs.