Remove deprecated code from Remote Access (#19077)
Closes #18831

### Summary of the issue:

Remote Access was using the deprecated `braille.filter_displaySize`
extension point.

It was also using `displaySize` instead of `displayDimensions.numCols`,
meaning that in the case that a user with a multi-line display was
controlling a computer with a single-line display that was wider than
the leader's display, braille from the follower would wrap at the width
of the leader display. While this would not result in information loss,
it was unexpected behaviour.

### Description of user facing changes:

In remote access sessions where one or more users have multi-line
braille displays connected, only the width of these displays will be
considered when finding the shared display width. For instance, if
leader connects with a DotPad (8x20 cells) and follower connects with a
Hims Brailleedge 40 (1x40 cells), the display width for the session will
be constrained to 20 cells.

### Description of developer facing changes:

None

### Description of development approach:

Replace usage of `braill.filter_displaySize` with
`braille.filter_displayDimensions` in `_remoteClient.session`. Also
renamed
`_remoteClient.localMachine.LocalMachine.handleFilterDisplaySize` to
`_handleFilterDisplayDimensions`.

Replace usages of `braille.handler.displaySize` with
`braille.handler.displayDimension.numCols` where appropriate in
`_remoteClient`.

### Testing strategy:

Created a session in which the leader was using a DotPad and the
follower was using a Brailleedge40.

### Known issues with pull request:

Users may desire the existing behaviour of all rows of multi-line
displays being considered.

We cannot support multi-line displays over Remote Access without
changing the protocol. NV Access does not have a clear policy on when we
can change the Remote Access Protocol and in what ways.