Fix Chrome system tests (#19121)
### Summary of the issue:

A recent update to chrome has introduced a "Ask Google about this page"
option in the omnibox (address bar). This option is sometimes spoken,
even though it isn't focused, when focusing the address bar. This causes
system tests which rely on Chrome to fail, as they are unable to
recognise that the address bar has been focused.
This breaks some tests in the Chrome, symbols, and image descriptions
system test suites.

### Description of user facing changes:

None.

### Description of developer facing changes:

System tests work again.

### Description of development approach:

When setting up a system test that uses Chrome, after sending `alt+d` to
focus the address bar, if the speech is not what we expect, try
reporting the focused object (`NVDA+tab`) to double-check.
This works because "Address and search bar" speech is cancelled by the
"Ask Google about this page" speech, so the system tests are unaware
where the focus has landed. By explicitly checking the current focus, we
work around this issue.

### Testing strategy:

Ran the Chrome system test "checkbox labelled by inner element" locally,
with and without this patch, to ensure that it was broken on my machine
without the patch, and fixed with it.
Ran in CI.

### Known issues with pull request:

This does not fix the underlying issue, which will still likely
inconvenience users. However, as this is stopping us being able to
create snapshot, beta, rc or release builds, this fix needed to be
prioritised.