Report number of items for lists in braille (#19119)
Closes #7455

### Summary of the issue:

In browse mode documents, the number of ítems for lists is not reported
in braille, just via speech output.

### Description of user facing changes:

The number of ítems Will be appended to lst for lists in browse mode
documents.

### Description of developer facing changes:

None

### Description of development approach:

Use the _childcontrolcount attribute of text info fields to get the
number of ítems of lists, and set the appropriate roleText when the
number of ítems of a list is available.

### Testing strategy:
Tested locally, checking that the number of ítems is reported for lists
in browse mode.

### Known issues with pull request:

In rare cases, the number of items can be reported via speech output,
not in braille, or in braille, not speech. Here is an example of a
webpage where the number of items is reported just via speech. I've
observed this just when the list contains 1 element.
In braille, in Spanish, in this webpage, pressing l to reach a list with
1 item, I read the word "fin" (end). This is not read via speech. Seems
a non standard control (like an empty list inside a menu):

https://duckduckgo.com/

I've seen, once, that the number of items was reported just in braille,
not speech. If I find the webpage I'll include it here.