In Outlook Classic, when navigating messages, ensure that wrong info is not displayed in braille (#19109)
Fixes #18993

### Summary of the issue:
When navigating the list of messages in Outlook Classic, the navigator
object is moved from the focus, showing unexpected information in
braille.

### Description of user facing changes:
The list of messages can be navigated in Outlook Classic, without
receiving unexpected information in braille.

### Description of developer facing changes:
None.

### Description of development approach:
In the `ContactEditField` class of the Outlook app module, call to
`api.setNavigatorObject` for the `event_valueChange`, has been modified
setting `isFocus` to `True`

### Testing strategy:
- Navigate the list of messages and chek that unexpected information is
not presented in braille.
- Compose messages using fields like To or CC, and check that side
effects are not produced.

### Known issues with pull request:
The navigator object is still moved, so that the reported navigator
object doesn't match the reported focus.