Excel legacy now reports Has note #19161
Fixes #18709
Follow-up of #11311
Summary of the issue:

In Excel (legacy), when entering a cell with a note, Excel reports "Has comment". This is confusing since there are new threaded comment types in newer Excel versions.

It was listed as a known issue of #11311.

It's worth noting that with Excel UIA, NVDA already reports "Has note" in this case.
Description of user facing changes:

When moving the focus on a cell with a note, Excel now reports "Has note". It is consistent with NVDA's interface and User Guide which mention note everywhere else.
Description of developer facing changes:

N/A
Description of development approach:

Use the new HASNOTE state introduced since then instead of the HASCOMMENT state, as done for UIA.