Fix language reporting dialects as not supported (#19116)
Link to issue number:

Fixes #18876
Summary of the issue:

NVDA reports as not supported dialects not available in synthesizer languages, even when a fallback language can be switched to.
Description of user facing changes:

When a dialect is not available for the current synthesizer, but NVDA can switch to a fallback language, that dialect won't be reported as not supported.
Description of developer facing changes:

None
Description of development approach:

In the languageIsSupportedfunction of the SynthDriver class, if a language is not supported, search for the availability of a fallback language, defined as mainLanguage-mainLanguage, for example, fr-fr.