Update Quick Start Guide (#19174)
### Summary of the issue:

The Quick Start Guide (QSG) contains the following issues:
* It mentions unassigned commands; unassigned commands should not be
mentioned in the Quick Start Guide since:
* if they are unassigned, they are probably not so common to deserve to
be in the QSG
* the QSG does not indicate how to assign a gesture to a command, which
makes mentioning unassigned commands useless in this guide.
* There is a paragraph on the synth settings ring commands in the QSG,
but the user would not even know what it is.
* Other unassigned commands are mentioned in the synth settings ring
paragraph of the QSG.

### Description of user facing changes:

* Removed unassigned "Report language" from the QSG.
* Added an initial description of the synth setting ring at the top of
its dedicated paragraph in the QSG.
* Removed the mention of unassigned commands to jumpt to the first/last
value in the synth ring commands since they are not assigned, for same
reason as above.

### Description of developer facing changes:

N/A

### Description of development approach:

N/A

### Testing strategy:

Check the generated User Guide.

### Known issues with pull request:

None