Note remote access limitations in secure mode section of User Guide (#19210)
### Summary of the issue:

The User Guide section 13.1 (Remote Access, Getting started) notes the
limitation for starting a new session in secure mode. This was not noted
in the secure mode section and may be missed by people looking for
information there.

### Description of user facing changes:

Added a line to the list of bullet points in 18.1 "Secure mode disables"
to note manually initiating a new remote access session.