Do not mention Notes in Doc formatting settings (#19212)
Fix-up of #11311

### Summary of the issue:

In #11311, it was taken into account that older Excel comments were
known as "notes" in newer interfaces, so NVDA's GUI and documentation
has been updatedt to be consistent with newer versions.

The "Comments" checkbox in the Document formatting settings panel has
been updated to "Notes and comments" as part of this work.

However, this checkbox does not control Excel's comments/notes
reporting; it only controls whether comments in text, e.g. in Word, are
reported.

### Description of user facing changes:

This checkbox has been renamed back to "Comments"

### Description of developer facing changes:

N/A

### Description of development approach:

N/A

### Testing strategy:

Manual check

### Known issues with pull request:

One could argue that this checkbox should instead remain "Notes and
comments" and should control Excel's Notes or comments.

That's not my approach. And allowing to control "has notes" cell's
property while "has formula" cannot be controlled would not be
consistent.