Update SCons to 4.10.1, introducing VS2026 support (#19226)
Summary of the issue:

SCons was updated recently. This is a necessary step to build NVDA with Visual Studio 2026, which is fully supported.
Description of user facing changes:

None
Description of developer facing changes:

Support for building with VS2026 is now there, however note that VS2022 is still the standard.
Description of development approach:

Updated SCons