User guide: MathCAT edits (#19234)
### Link to issue number:
None

### Summary of the issue:
Edited the MathCAT section of the user guide, both the overview and
settings sections for clarity and organization.

### Description of user facing changes:
User guide content edits.

### Description of developer facing changes:
The following edits were made:

* Typical use is now a level 4 heading
* ETS acronym expanded (Educational Testing Service)

### Description of development approach:
Edited and proofread to make sure content is clearer.

### Testing strategy:
None (manual proofreading is recommended)

### Known issues with pull request:
None

### Code Review Checklist:

<!--
This checklist is a reminder of things commonly forgotten in a new PR.
Authors, please do a self-review of this pull-request.
Check items to confirm you have thought about the relevance of the item.
Where items are missing (eg unit / system tests), please explain in the
PR.
To check an item `- [ ]` becomes `- [x]`, note spacing.
You can also check the checkboxes after the PR is created.
A detailed explanation of this checklist is available here:

https://github.com/nvaccess/nvda/blob/master/projectDocs/dev/githubPullRequestTemplateExplanationAndExamples.md#code-review-checklist
-->

- [x] Documentation:
  - Change log entry
  - User Documentation
  - Developer / Technical Documentation
  - Context sensitive help for GUI changes
- [x] Testing:
  - Unit tests
  - System (end to end) tests
  - Manual testing
- [x] UX of all users considered:
  - Speech
  - Braille
  - Low Vision
  - Different web browsers
  - Localization in other languages / culture than English
- [x] API is compatible with existing add-ons.
- [x] Security precautions taken.

---------

Co-authored-by: Sean Budd <seanbudd123@gmail.com>