Update user_docs/en/userGuide.md
I think because we have called the feature "Screen Curtain" (and it's capitalised like feature names we've used) - but the words themselves are common and could simply be written in a sentence which would make sense even if we didn't have such a feature (and in that case you would use "the").  If we change this, we should probably change other references to match.

Co-authored-by: Sean Budd <sean@nvaccess.org>